requirements.txt
backports.entry-points-selectable==1.1.0
cfgv==3.3.0
click==8.0.1
distlib==0.3.2
filelock==3.0.12
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
identify==2.2.11
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
nodeenv==1.6.0
platformdirs==2.0.2
pre-commit==2.13.0
psycopg2==2.7.7
pycparser==2.20
python-dotenv==0.18.0
PyYAML==5.4.1
six==1.16.0
SQLAlchemy==1.4.22
toml==0.10.2
typing-extensions==3.10.0.0
virtualenv==20.6.0
Werkzeug==2.0.1
zipp==3.5.0
